Move bin target and related functions to web module
