+++ src/web/mod.rs
@@ -1,5 +1,6 @@
 //! Utilities for building and running the app in the browser.
 
+pub(crate) mod bin_target;
 pub(crate) mod build;
 pub(crate) mod bundle;
 pub(crate) mod profiles;
